presents the new exact traveling wave solutions of fourth order (1+1)- dimensional Boussinesq equation.
Proposed a new exponential expansion method and
apply to undertake this study. This study further shows some solitary
and periodic waves
graphically. This paper also shows that the novel exponential expansion
method is easily
applicable and powerful mathematical tool in the symbolic computational
approach in the
field of mathematical physics and engineering. The exact solutions of
this equation play a
vital role for describing different types of wave propagation in any
varied natural instances.